[PAY-27, C-753] Remove tipping feature flag, Fix mutuals icon
