+++ packages/mobile/src/screens/profile-screen/ProfileHeader/Bio.tsx
