+++ packages/mobile/src/screens/profile-screen/ProfileHeader/CollapsedSection.tsx
