+++ packages/mobile/src/screens/profile-screen/ProfileHeader/ExpandHeaderToggleButton.tsx
