+++ packages/mobile/src/screens/profile-screen/ProfileHeader/ExpandedSection.tsx
