+++ packages/mobile/src/screens/profile-screen/ProfileHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './ProfileHeader'
