--- packages/mobile/src/screens/profile-screen/ProfileHeaderV2/index.ts
@@ -1,1 +1,0 @@
-export * from './ProfileHeaderV2'
