+++ packages/mobile/src/screens/profile-screen/ProfileScreen.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from 'react'
 
-import { Status, ShareSource, FeatureFlags } from '@audius/common'
+import { Status, ShareSource } from '@audius/common'
 import { PortalHost } from '@gorhom/portal'
 import { getUserId } from 'audius-client/src/common/store/account/selectors'
 import { fetchProfile } from 'audius-client/src/common/store/pages/profile/actions'
@@ -15,7 +15,6 @@ import { IconButton, Screen } from 'app/components/core'
 import { useDispatchWeb } from 'app/hooks/useDispatchWeb'
 import { useNavigation } from 'app/hooks/useNavigation'
 import { usePopToTopOnDrawerOpen } from 'app/hooks/usePopToTopOnDrawerOpen'
-import { useFeatureFlag } from 'app/hooks/useRemoteConfig'
 import { useSelectorWeb } from 'app/hooks/useSelectorWeb'
 import { TopBarIconButton } from 'app/screens/app-screen'
 import { makeStyles } from 'app/styles/makeStyles'
@@ -24,7 +23,6 @@ import { useThemeColors } from 'app/utils/theme'
 import type { ProfileTabScreenParamList } from '../app-screen/ProfileTabScreen'
 
 import { ProfileHeader } from './ProfileHeader'
-import { ProfileHeaderV2 } from './ProfileHeaderV2'
 import { ProfileTabNavigator } from './ProfileTabNavigator'
 import { useSelectProfileRoot } from './selectors'
 
@@ -56,9 +54,6 @@ export const ProfileScreen = () => {
   const [isRefreshing, setIsRefreshing] = useState(false)
   const { neutralLight4, accentOrange } = useThemeColors()
   const navigation = useNavigation<ProfileTabScreenParamList>()
-  const { isEnabled: isTippingEnabled } = useFeatureFlag(
-    FeatureFlags.TIPPING_ENABLED
-  )
 
   const handlePressSettings = useCallback(() => {
     navigation.push({
@@ -129,13 +124,10 @@ export const ProfileScreen = () => {
 
   const scrollY = useRef(new Animated.Value(0)).current
 
-  const renderHeader = useCallback(() => {
-    return isTippingEnabled ? (
-      <ProfileHeaderV2 scrollY={scrollY} />
-    ) : (
-      <ProfileHeader scrollY={scrollY} />
-    )
-  }, [isTippingEnabled, scrollY])
+  const renderHeader = useCallback(
+    () => <ProfileHeader scrollY={scrollY} />,
+    [scrollY]
+  )
 
   return (
     <Screen topbarLeft={topbarLeft} topbarRight={topbarRight}>
