+++ packages/mobile/src/screens/user-list-screen/UserList.tsx
@@ -1,10 +1,8 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 
-import { FeatureFlags } from '@audius/common'
 import type { ID, User } from '@audius/common'
 import { useFocusEffect, useIsFocused } from '@react-navigation/native'
 import type { CommonState } from 'audius-client/src/common/store'
-import { getUserId } from 'audius-client/src/common/store/account/selectors'
 import { getUsers } from 'audius-client/src/common/store/cache/users/selectors'
 import {
   loadMore,
@@ -19,11 +17,9 @@ import type { Selector } from 'react-redux'
 import { Divider, FlatList } from 'app/components/core'
 import LoadingSpinner from 'app/components/loading-spinner'
 import { useDispatchWeb } from 'app/hooks/useDispatchWeb'
-import { useFeatureFlag } from 'app/hooks/useRemoteConfig'
 import { isEqual, useSelectorWeb } from 'app/hooks/useSelectorWeb'
 import { makeStyles } from 'app/styles'
 
-import { UserChip } from './UserChip'
 import { UserListItem } from './UserListItem'
 
 const useStyles = makeStyles(({ spacing }) => ({
@@ -73,7 +69,6 @@ export const UserList = (props: UserListProps) => {
     tag
   })
   const optimisticUserIds: ID[] = useSelectorWeb(getOptimisticUserIds)
-  const currentUserId = useSelectorWeb(getUserId)
   const usersMap = useSelectorWeb(
     (state) => getUsers(state, { ids: optimisticUserIds }),
     isEqual
@@ -84,10 +79,6 @@ export const UserList = (props: UserListProps) => {
         .map((id) => usersMap[id])
         .filter((user) => user && !user.is_deactivated),
     [usersMap, optimisticUserIds]
-  )
-
-  const { isEnabled: isTippingEnabled } = useFeatureFlag(
-    FeatureFlags.TIPPING_ENABLED
   )
 
   useFocusEffect(
@@ -146,15 +137,9 @@ export const UserList = (props: UserListProps) => {
     <FlatList
       style={styles.list}
       data={data}
-      renderItem={({ item }) =>
-        isTippingEnabled ? (
-          <UserListItem user={item} tag={tag} />
-        ) : (
-          <UserChip user={item} currentUserId={currentUserId} />
-        )
-      }
+      renderItem={({ item }) => <UserListItem user={item} tag={tag} />}
       keyExtractor={(item) => item.user_id.toString()}
-      ItemSeparatorComponent={isTippingEnabled ? Divider : undefined}
+      ItemSeparatorComponent={Divider}
       onEndReached={handleEndReached}
       ListFooterComponent={loading || isRefreshing ? loadingSpinner : footer}
     />
