+++ packages/mobile/src/screens/user-list-screen/UserListScreen.tsx
@@ -1,11 +1,9 @@
 import type { ComponentType, ReactElement, ReactNode } from 'react'
 import { useCallback } from 'react'
 
-import { FeatureFlags } from '@audius/common'
 import type { SvgProps } from 'react-native-svg'
 
 import { Screen } from 'app/components/core'
-import { useFeatureFlag } from 'app/hooks/useRemoteConfig'
 
 import { UserListTitle } from './UserListTitle'
 
@@ -17,9 +15,6 @@ type UserListScreenProps = {
 
 export const UserListScreen = (props: UserListScreenProps) => {
   const { title, titleIcon, children } = props
-  const { isEnabled: isTippingEnabled } = useFeatureFlag(
-    FeatureFlags.TIPPING_ENABLED
-  )
 
   const headerTitle = useCallback(() => {
     if (!titleIcon) {
@@ -29,12 +24,7 @@ export const UserListScreen = (props: UserListScreenProps) => {
   }, [titleIcon, title])
 
   return (
-    <Screen
-      variant='white'
-      title={isTippingEnabled ? undefined : title}
-      headerTitle={isTippingEnabled ? headerTitle : undefined}
-      topbarRight={null}
-    >
+    <Screen variant='white' headerTitle={headerTitle} topbarRight={null}>
       {children}
     </Screen>
   )
