+++ packages/web/src/pages/profile-page/components/desktop/ProfileLeftNav.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react'
 
-import { ID, Name, FeatureFlags } from '@audius/common'
+import { ID, Name } from '@audius/common'
 import cn from 'classnames'
 import { animated } from 'react-spring'
 
@@ -17,7 +17,6 @@ import ProfilePageBadge from 'components/user-badges/ProfilePageBadge'
 import { Type } from 'pages/profile-page/components/SocialLink'
 import SocialLinkInput from 'pages/profile-page/components/SocialLinkInput'
 import { ProfileTags } from 'pages/profile-page/components/desktop/ProfileTags'
-import { getFeatureEnabled } from 'services/remote-config/featureFlagHelpers'
 import { make, useRecord } from 'store/analytics/actions'
 import { UPLOAD_PAGE } from 'utils/route'
 
@@ -94,7 +93,6 @@ export const ProfileLeftNav = (props: ProfileLeftNavProps) => {
   } = props
 
   const record = useRecord()
-  const isTippingEnabled = getFeatureEnabled(FeatureFlags.TIPPING_ENABLED)
   const accountUser = useSelector(getAccountUser)
 
   const onClickUploadChip = useCallback(() => {
@@ -201,12 +199,11 @@ export const ProfileLeftNav = (props: ProfileLeftNavProps) => {
           instagramHandle={instagramHandle}
           tikTokHandle={tikTokHandle}
         />
-        {isTippingEnabled &&
-        (!accountUser || accountUser.user_id !== userId) ? (
+        {!accountUser || accountUser.user_id !== userId ? (
           <OpacityTransition render={renderTipAudioButton} />
         ) : null}
-        {isTippingEnabled && <SupportingList />}
-        {isTippingEnabled && <TopSupporters />}
+        <SupportingList />
+        <TopSupporters />
         {isArtist ? <ProfileTags goToRoute={goToRoute} tags={tags} /> : null}
         <ProfileMutuals />
         {isOwner && !isArtist && (
